add some details for call centre

+++ bot/index.js
@@ -5,9 +5,10 @@ const mongoose = require("mongoose");
 const Products = require("../models/Products.js");
 const LocalStorage = require("../models/Localstorage.js");
 const Orders = require("../models/Orders.js");
+const Users = require("../models/Users.js");
 require("dotenv").config();
 
-const token = "7496758951:AAHPGeexlNH8meAV8G8AKHxE7HDr-bQFL4g";
+const token = process.env.BOT_TOKEN;
 
 const bot = new TelegramBot(token, {polling: true});
 
@@ -42,9 +43,16 @@ bot.on("message", async (msg) => {
 
 		preorder.products.forEach((item) => {
 			const product = item.product;
-			const productDetails = `- ${product.name_uz} (${product.brand.name}), ${product.category.name_uz}, ${item.product.price} UZS\n`;
+			const productDetails = `- ${product.name_uz} (${product.brand.name}), ${
+				product.category.name_uz
+			}, ${
+				item.product.sale.isSale ? item.product.sale.price : item.product.price
+			} so'm (${item.quantity} ta)\n`;
 			productList += productDetails;
-			totalPrice += item.product.price;
+			totalPrice +=
+				(item.product.sale.isSale
+					? item.product.sale.price
+					: item.product.price) * item.quantity;
 		});
 
 		const create = await LocalStorage.create({
@@ -58,7 +66,7 @@ Qabul qiluvchi: ${preorder.name} ${preorder.surname}
 Telefon raqami: ${preorder.phone_number}
 Mahsulotlar:
 ${productList}
-Umumiy narx: ${totalPrice} UZS
+Umumiy narx: ${totalPrice} so'm
 
 Yetkaziladigan Manzilni Yuboring`;
 
@@ -77,22 +85,52 @@ Yetkaziladigan Manzilni Yuboring`;
 			},
 		});
 	}
+	if (text == "Mening Buyurtmalarim") {
+		const user = await Users.findOne({
+			"telegram.id": chatId,
+		});
+		const orders = await Orders.find({
+			userId: new mongoose.Types.ObjectId(user._id),
+		});
+		console.log(orders);
+		bot.sendMessage(chatId, "Buyurtmalaringiz");
+	}
 });
 
 bot.on("location", async (msg) => {
 	const chatId = msg.chat.id;
-	const userStorage = await LocalStorage.findOne({
-		name: chatId,
-	});
+	const userStorage = await LocalStorage.findOne({name: chatId});
+
 	if (userStorage && stripos(userStorage.value, "location||")) {
-		const {latitude, longitude} = msg.location;
+		console.log(msg.location);
 		const id = userStorage.value.split("||")[1];
 		const preorder = await PreOrders.findById(new mongoose.Types.ObjectId(id));
-		preorder.delivery.address.longitude = longitude;
-		preorder.delivery.address.latitude = latitude;
+
+		if (!preorder) {
+			await bot.sendMessage(
+				chatId,
+				"Buyurtma topilmadi. Iltimos, yana urinib ko'ring.",
+			);
+			return;
+		}
+
+		preorder.delivery.address.latitude = msg.location.latitude;
+		preorder.delivery.address.longitude = msg.location.latitude;
 		await preorder.save();
 		await userStorage.deleteOne();
+
+		let user = await Users.findOne({"telegram.id": chatId});
+		if (!user) {
+			user = await Users.create({
+				name: preorder.name,
+				surname: preorder.surname,
+				telegram: {id: chatId},
+			});
+			await user.save();
+		}
+
 		const newOrder = await Orders.create({
+			userId: user._id,
 			delivery: {
 				date: preorder.delivery.date,
 				address: {
@@ -105,25 +143,56 @@ bot.on("location", async (msg) => {
 			products: preorder.products,
 			comment: preorder.delivery.comment || "",
 		});
+
+		const uzumUrl =
+			"https://www.apelsin.uz/open-service?serviceId=498616071&order_id=" +
+			newOrder.order_id;
+
+		let totalAmount = 0;
+		for (const product of newOrder.products) {
+			const productDoc = await Products.findById(product.product);
+			if (!productDoc) {
+				await bot.sendMessage(
+					chatId,
+					"Mahsulot topilmadi. Iltimos, yana urinib ko'ring.",
+				);
+				return;
+			}
+			const price = productDoc.sale.isSale
+				? productDoc.sale.price
+				: productDoc.price;
+			const subtotal = price * product.quantity;
+			totalAmount += subtotal;
+		}
+
+		const stringToEncode = `m=663b1ac0fe41a3907df8f595;ac.order_id=${
+			newOrder.order_id
+		};a=${totalAmount * 100}`;
+		const base64EncodedString = Buffer.from(stringToEncode).toString("base64");
+		const paymeUrl = "https://checkout.paycom.uz/" + base64EncodedString;
+		const clickUrl = `https://my.click.uz/services/pay?service_id=33923&merchant_id=25959&amount=${totalAmount}&transaction_param=${newOrder.order_id}`;
+
 		await newOrder.save();
-		await bot.sendMessage(
-			chatId,
-			`
-			Joylashuvingiz qabul qilindi
-To'lovni Amalga oshiring`,
-			{
-				reply_markup: {
-					resize_keyboard: true,
-					inline_keyboard: [
-						[
-							{
-								text: "Joylashuvni Yuborish",
-								url: "",
-							},
-						],
-					],
-				},
+		await preorder.deleteOne();
+
+		await bot.sendMessage(chatId, "Joylashuvingiz qabul qilindi", {
+			reply_markup: {
+				resize_keyboard: true,
+				keyboard: [[{text: "Mening Buyurtmalarim"}]],
 			},
-		);
+		});
+
+		await bot.sendMessage(chatId, "To'lovni Amalga oshiring", {
+			reply_markup: {
+				resize_keyboard: true,
+				inline_keyboard: [
+					[
+						{text: "Uzum", url: uzumUrl},
+						{text: "Payme", url: paymeUrl},
+					],
+					[{text: "Click", url: clickUrl}],
+				],
+			},
+		});
 	}
 });
